Return player object vs. _id from method.
This alleviates the need to do an extra round-trip fetch when the update succeeds.
When the server method fails, currently the reducer updates the state, but it feels like the server should send down the player object on error as well to be consistent with the way succeed flow works.